dashboard page chnage into financial and cards update from wesite and detail page chnage gallery image grid
