+++ public/frontend-assets/assets/css/custom.css
@@ -25,7 +25,16 @@ h1, h2, h3, h4, h5, h6 {
   background-color: var(--text-blue) !important;
   color: white !important;
 }
-
+.custom-badge-2 {
+    background-color: var(--hover-clr);
+    color: var(--text-orange);
+    border-radius: 50px;
+    padding: 8px 16px;
+    font-size: 14px;
+    display: inline-flex;
+    align-items: center;
+    font-weight: 500;
+}
 * {
   margin: 0;
   padding: 0;
@@ -4011,4 +4020,98 @@ margin: 5px 0;
 }
 
 /* end css for smart dropdown design */
+  .car-gallery {
+        display: flex;
+        gap: 16px;
+    }
 
+    .main-image {
+        flex: 2;
+        height: 520px;
+        overflow: hidden;
+        border-top-left-radius: 16px;
+        border-bottom-left-radius: 16px;
+        box-shadow: 0 4px 12px rgba(0,0,0,0.1);
+    }
+
+    .main-image img {
+        width: 100%;
+        height: 100%;
+        object-fit: fill;
+        object-position: center;
+        display: block;
+    }
+
+    .side-images {
+        flex: 1;
+        display: grid;
+        grid-template-columns: 1fr 1fr;
+        grid-template-rows: 1fr 1fr;
+        gap: 8px;
+        height: 520px;
+    }
+
+    .side-images .image-wrapper {
+        width: 100%;
+        height: 100%;
+        overflow: hidden;
+    }
+
+    .side-images img {
+        width: 100%;
+        height: 100%;
+        object-fit: fill;
+        object-position: center;
+        display: block;
+    }
+
+    /* Rounded corners only for specific side images */
+    .side-images .image-wrapper:nth-child(2) {
+        border-top-right-radius: 16px;
+    }
+
+    .side-images .image-wrapper:nth-child(4) {
+        border-bottom-right-radius: 16px;
+    }
+     /* Responsive Layout for Tablets and Smaller Screens */
+    @media (max-width: 992px) {
+        .car-gallery {
+            flex-direction: column;
+        }
+
+        .main-image,
+        .side-images {
+            height: auto;
+            border-radius: 0;
+        }
+
+        .side-images {
+            grid-template-columns: repeat(2, 1fr);
+            grid-template-rows: auto;
+        }
+
+        .side-images .image-wrapper:nth-child(2),
+        .side-images .image-wrapper:nth-child(4) {
+            border-radius: 0;
+        }
+    }
+
+    /* Mobile Phones */
+    @media (max-width: 576px) {
+        .side-images {
+            grid-template-columns: 1fr;
+            grid-template-rows: repeat(4, auto);
+        }
+
+        .car-gallery {
+            gap: 12px;
+        }
+
+        .main-image {
+            height: 300px;
+        }
+
+        .side-images .image-wrapper {
+            height: 180px;
+        }
+    }
