+++ resources/lang/ar/messages.php
@@ -667,4 +667,22 @@ return [
 'per_hour'   => 'لكل ساعة',
 'user' => 'المستخدم',
 'no_booking' => 'لم يتم العثور على حجوزات.',
+'featured_vehicle' => 'السيارة المميزة',
+'newArrival' => 'وصل حديثًا',
+'financial' => 'مالي',
+'order_confirm' => 'الطلبات المؤكدة',
+'total_revenue' => 'إجمالي الإيرادات',
+'order_complete' => 'الطلبات المكتملة',
+'order_cancel' => 'الطلبات الملغاة',
+'order_pending' => 'الطلبات المعلقة',
+'pending_revenue' => 'الإيرادات المعلقة',
+'order_status' => 'حالة الطلبات',
+'week' => 'هذا الأسبوع',
+'month' => 'هذا الشهر',
+'last_month' => 'الشهر الماضي',
+'this_year' => 'هذا العام',
+'booking_overview' => 'نظرة عامة على الحجوزات',
+'reminders' => 'التذكيرات',
+'earning_summary' => 'ملخص الأرباح',
+
 ];
