+++ resources/lang/en/messages.php
@@ -647,7 +647,7 @@ return [
 'user_info'=>'User Info',
 'company_info'=>'Compnay Info',
 'company_profile'=>'Company Profile',
-'company_name'=>'Company Name',
+'company_name'=>'Company',
 'company_email'=>'Company Email',
 'company_phone'=>'Company Phone',
 'company_website'=>'Company Website',
@@ -667,4 +667,22 @@ return [
 'per_hour'      => 'Per Hour',
 'user'=>'User',
 'no_booking'=>'No bookings found.',
+'featured_vehicle'=>'Featured Vehicle',
+'newArrival'=>'New Arrival',
+'financial'=>'Financial',
+'order_confirm'=>'Orders Confirmed',
+'total_revenue'=>'Total Revenue',
+'order_complete'=>'Orders Completed',
+'order_cancel'=>'Orders Cancelled',
+'order_pending'=>'Orders Pending',
+'pending_revenue'=>'Pending Revenue',
+'order_status'=>'Orders Status',
+'week'=>'This Week',
+'month'=>'This Month',
+'last_month'=>'Last Month',
+'this_year'=>'This Year',
+'booking_overview'=>'Bookings Overview',
+'reminders'=>'Reminders',
+'earning_summary'=>'Earnings Summary',
+''=>'',
 ];
