+++ resources/lang/nl/messages.php
@@ -662,4 +662,22 @@ return [
 'per_hour'   => 'Per uur',
 'user' => 'Gebruiker',
 'no_booking' => 'Geen boekingen gevonden.',
+'featured_vehicle' => 'Uitgelicht voertuig',
+'newArrival' => 'Nieuwe aankomst',
+'financial' => 'Financieel',
+'order_confirm' => 'Bevestigde bestellingen',
+'total_revenue' => 'Totale opbrengst',
+'order_complete' => 'Voltooide bestellingen',
+'order_cancel' => 'Geannuleerde bestellingen',
+'order_pending' => 'In afwachting zijnde bestellingen',
+'pending_revenue' => 'In afwachting zijnde opbrengst',
+'order_status' => 'Bestelstatus',
+'week' => 'Deze week',
+'month' => 'Deze maand',
+'last_month' => 'Vorige maand',
+'this_year' => 'Dit jaar',
+'booking_overview' => 'Boekingsoverzicht',
+'reminders' => 'Herinneringen',
+'earning_summary' => 'Inkomstenoverzicht',
+
 ];
