+++ resources/views/website/car-listing/car-listing-filters/car-list.blade.php
@@ -1,5 +1,65 @@
 @foreach($cars as $car)
-<div class="col-sm-6 col-md-6 col-lg-4 mb-5 car-item">
+<div class="col-sm-6 col-md-6 col-lg-6 mb-4">
+    <div class="custom-card2 p-3 shadow-sm rounded">
+        <a href="{{ url('/cardetail/' . $car->id) }}" class="link text-decoration-none">
+            <img src="{{ asset('storage/' . $car->thumbnail) }}" class="custom-card-img img-fluid mb-2 rounded" alt="Car Image">
+        </a>
+        <div class="card-content">
+            <div class="d-flex justify-content-between align-items-start mb-2">
+                <div>
+                    <h5 class="card_orange_clr mb-1" style="font-weight: 600;">
+                        {{ $car->car_models->car_brands->name ?? '' }}
+                        {{ $car->car_models->name ?? '' }}
+                        {{ $car->year ?? '' }}
+                    </h5>
+                    <h6 class="text-muted mb-0" style="font-size: 12px;">
+                        {{ __('messages.engine') }} {{ $car->engine_size }}
+                        {{ __('messages.cc') }} | {{ ucfirst($car->transmission) }} | {{ $car->fuel_type }}
+                    </h6>
+                </div>
+                <div class="ms-2">
+                    @php
+                        $companyProfile = $car->users->companies->company_profile ?? null;
+                    @endphp
+                    <img src="{{ asset($companyProfile ?? 'frontend-assets/assets/customeruser.png') }}" alt="Company Logo"
+                         width="40" height="40" class="rounded-circle" style="object-fit: cover;">
+                </div>
+            </div>
+
+            <hr class="my-2">
+
+            <div class="d-flex justify-content-between text-center mb-2">
+                <div class="flex-fill">
+                    <h6 class="mb-1" style="font-size: 13px;">{{ __('messages.mileage') }}</h6>
+                    <p class="mb-0" style="font-size: 14px;"><i class="fas fa-tachometer-alt me-1"></i>{{ $car->mileage }}</p>
+                </div>
+                <div class="flex-fill">
+                    <h6 class="mb-1" style="font-size: 13px;">{{ __('messages.deposit') }}</h6>
+                    <p class="mb-0" style="font-size: 14px;">{{ convertPrice($car->advance_deposit, 0) }}</p>
+                </div>
+                <div class="flex-fill">
+                    <h6 class="mb-1" style="font-size: 13px;">{{ __('messages.min') }}</h6>
+                    <p class="mb-0" style="font-size: 14px;"><i class="fa-solid fa-user me-1"></i>{{ $car->min_age }}</p>
+                </div>
+            </div>
+
+            <hr class="my-2">
+
+            <div class="d-flex justify-content-between align-items-center bg-light rounded px-2 py-1">
+                <h6 class="card_orange_clr mb-0">{{ convertPrice($car->rent, 0) }}/{{ __('messages.Day') }}</h6>
+                @if (auth()->check())
+                    <button class="book-btn" data-carid="{{ $car->id }}" id="car-booking-btn">{{ __('messages.Book') }}</button>
+                @else
+                    <button class="book-btn" data-bs-toggle="modal" data-bs-target="#loginModal">{{ __('messages.Book') }}</button>
+                @endif
+            </div>
+        </div>
+    </div>
+</div>
+@endforeach
+
+{{-- old design --}}
+{{-- <div class="col-sm-6 col-md-6 col-lg-4 mb-5 car-item">
     <div class="car-listing-card">
         <a href="{{ url('/cardetail/' . $car->id) }}" class="link">
         <img src="{{ asset('storage/' . $car->thumbnail) }}" alt="{{ $car->car_models->name ?? 'Car Image' }}" class="listing-car-image mb-2">
@@ -35,10 +95,6 @@
                 </div>
             </div>
         </a>
-            {{-- <button onclick="window.location.href='{{ url('/cardetail/' . $car->id) }}'" class="details-btn rounded-pill">
-                Details
-            </button>             --}}
-        </div>
     </div>
-</div>
-@endforeach
+    </div>
+</div> --}}
